fix: reemplazar authorizeResource con authorize manual en CategoriaController

+++ app/Http/Controllers/CategoriaController.php
@@ -16,7 +16,6 @@ class CategoriaController extends Controller
     public function __construct(
         private CategoriaService $categoriaService
     ) {
-        $this->authorizeResource(Categoria::class, 'categoria');
     }
 
     /**
@@ -24,6 +23,8 @@ class CategoriaController extends Controller
      */
     public function index(): Response
     {
+        $this->authorize('viewAny', Categoria::class);
+        
         $categorias = $this->categoriaService->obtenerCategoriasUsuario();
 
         return Inertia::render('Categorias/Index', [
@@ -36,6 +37,8 @@ class CategoriaController extends Controller
      */
     public function store(CategoriaData $data): RedirectResponse
     {
+        $this->authorize('create', Categoria::class);
+        
         try {
             $this->categoriaService->crearCategoria($data);
 
@@ -50,6 +53,8 @@ class CategoriaController extends Controller
      */
     public function update(CategoriaData $data, Categoria $categoria): RedirectResponse
     {
+        $this->authorize('update', $categoria);
+        
         try {
             $this->categoriaService->actualizarCategoria($categoria, $data);
 
@@ -69,6 +74,8 @@ class CategoriaController extends Controller
      */
     public function destroy(Categoria $categoria): RedirectResponse
     {
+        $this->authorize('delete', $categoria);
+        
         try {
             $resultado = $this->categoriaService->eliminarCategoria($categoria);
